+++ src/Record/HlsMakerImp.h
@@ -54,7 +54,7 @@ protected:
     void onDelSegment(uint64_t index) override;
     void onWriteSegment(const char *data, size_t len) override;
     void onWriteHls(const std::string &data) override;
-    void onFlushLastSegment(uint32_t duration_ms) override;
+    void onFlushLastSegment(uint64_t duration_ms) override;
 
 private:
     std::shared_ptr<FILE> makeFile(const std::string &file,bool setbuf = false);
