+++ src/TS/TSMediaSourceMuxer.h
@@ -66,7 +66,7 @@ public:
     }
 
 protected:
-    void onWrite(std::shared_ptr<toolkit::Buffer> buffer, uint32_t timestamp, bool key_pos) override {
+    void onWrite(std::shared_ptr<toolkit::Buffer> buffer, uint64_t timestamp, bool key_pos) override {
         if (!buffer) {
             return;
         }
